+++ Web_Api2/WebApi_Example/WebApi_Example_Domain/Repository/Bookings/IBookingRepository.cs
@@ -6,7 +6,7 @@ namespace WebApi_Example_Domain.Repository
 {
     public interface IBookingRepository
     {
-        Task<bool> AddBooking(Bookings bookings, int roomID);
+        Task<bool> AddBooking(Bookings bookings, short roomID);
         Task<IEnumerable<Bookings>> GetBookings();
         Task<Bookings> GetBooking(int id);
         Task<bool> UpdateBooking(int id);
